Finished the DBImpl of Tracker, tests left to be done.

+++ src/client_info.cc
@@ -1,6 +1,6 @@
 #include "client_info.h"
 
-ClientInfo::ClientInfo(const string& clientId, const string& ip, UInt16 messagePort = 9025)
+ClientInfo::ClientInfo(const string& clientId, const string& ip, int messagePort = 9025)
     :clientId_(clientId),
     ip_(ip),
     messagePort_(messagePort)
